Further separation, flight search WIP
